+++ R/side-coord-cartesian.R
@@ -5,9 +5,9 @@
 NULL
 ### INCLUDE END
 
-#'@rdname ggside-ggproto-coord
-#'@title Coord Compatible with ggside
-#'@description
+#' @rdname ggside-ggproto-coord
+#' @title Coord Compatible with ggside
+#' @description
 #' S3 class that converts old Coord into one that
 #' is compatible with ggside. Can also update
 #' ggside on the object. Typically, the new ggproto
@@ -18,22 +18,23 @@ ggside_coord <- function(coord) UseMethod("ggside_coord")
 
 #' @rdname ggside-ggproto-coord
 #' @export
-ggside_coord.default <- function(coord){
+ggside_coord.default <- function(coord) {
   abort(glue("No known method to make {class(coord)[1]} ggside friendly"))
 }
 
 #' @rdname ggside-ggproto-coord
 #' @export
-ggside_coord.CoordCartesian <- function(coord){
+ggside_coord.CoordCartesian <- function(coord) {
   # insure classes that inherit from CoordCartesian fail
   # if there is no S3 method called.
-  if (class(coord)[1L]!="CoordCartesian") abort(glue("No known method to make {class(coord)[1]} ggside friendly"))
+  if (class(coord)[1L] != "CoordCartesian") abort(glue("No known method to make {class(coord)[1]} ggside friendly"))
   ggplot2::ggproto("CoordSide",
-                   CoordSideCartesian,
-                   limits = coord$limits,
-                   expand = coord$expand,
-                   default = coord$default,
-                   clip = coord$clip)
+    CoordSideCartesian,
+    limits = coord$limits,
+    expand = coord$expand,
+    default = coord$default,
+    clip = coord$clip
+  )
 }
 
 #' @rdname ggside-ggproto-coord
@@ -48,14 +49,13 @@ CoordSideCartesian <- ggplot2::ggproto(
   setup_panel_params = new_ggproto_fun(
     ggplot2::CoordCartesian$setup_panel_params,
     {
-
       if (!is.null(old_y_lim <- self$limits$y) &
-           is.ggside_scale(scale_y)) {
+        is_ggside_scale(scale_y)) {
         self$limits$y <- NULL
         out <- call_parent_method
         self$limits$y <- old_y_lim
       } else if (!is.null(old_x_lim <- self$limits$x) &
-          is.ggside_scale(scale_x)) {
+        is_ggside_scale(scale_x)) {
         self$limits$x <- NULL
         out <- call_parent_method
         self$limits$x <- old_x_lim
@@ -63,7 +63,8 @@ CoordSideCartesian <- ggplot2::ggproto(
         out <- call_parent_method
       }
       out
-    }),
+    }
+  ),
   render_bg = function(panel_params, theme) {
     panel_type <- panel_params$ggside_panel_type
     if (is.element(panel_type, c("x", "y"))) {
@@ -86,26 +87,33 @@ CoordSideCartesian <- ggplot2::ggproto(
     }
   },
   render_fg = ggside_render_fg,
-  render_axis_h = function (panel_params, theme) {
+  render_axis_h = function(panel_params, theme) {
     panel_type <- panel_params$ggside_panel_type
-    if (panel_type=="y") {
-      list(top = ggside_panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
-           bottom = ggside_panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom))
+    if (panel_type == "y") {
+      list(
+        top = ggside_panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
+        bottom = ggside_panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom)
+      )
     } else {
-      list(top = panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
-           bottom = panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom))
-
+      list(
+        top = panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
+        bottom = panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom)
+      )
     }
   },
-  render_axis_v = function (panel_params, theme) {
+  render_axis_v = function(panel_params, theme) {
     panel_type <- panel_params$ggside_panel_type
 
-    if (panel_type=="x") {
-      list(left = ggside_panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
-           right = ggside_panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right))
+    if (panel_type == "x") {
+      list(
+        left = ggside_panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
+        right = ggside_panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right)
+      )
     } else {
-      list(left = panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
-           right = panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right))
+      list(
+        left = panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
+        right = panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right)
+      )
     }
   }
 )
@@ -115,11 +123,12 @@ CoordSideCartesian <- ggplot2::ggproto(
 #' @export
 ggside_coord.CoordTrans <- function(coord) {
   ggplot2::ggproto("CoordSide",
-                   CoordSideTrans,
-                   trans = coord$trans,
-                   limits = coord$limits,
-                   expand = coord$expand,
-                   clip = coord$clip)
+    CoordSideTrans,
+    trans = coord$trans,
+    limits = coord$limits,
+    expand = coord$expand,
+    clip = coord$clip
+  )
 }
 
 CoordSideTrans <- ggplot2::ggproto(
@@ -129,12 +138,12 @@ CoordSideTrans <- ggplot2::ggproto(
     ggplot2::CoordTrans$setup_panel_params,
     {
       if (!is.null(old_y_lim <- self$limits$y) &
-          is.ggside_scale(scale_y)) {
+        is_ggside_scale(scale_y)) {
         self$limits$y <- NULL
         out <- call_parent_method
         self$limits$y <- old_y_lim
       } else if (!is.null(old_x_lim <- self$limits$x) &
-                 is.ggside_scale(scale_x)) {
+        is_ggside_scale(scale_x)) {
         self$limits$x <- NULL
         out <- call_parent_method
         self$limits$x <- old_x_lim
@@ -142,9 +151,10 @@ CoordSideTrans <- ggplot2::ggproto(
         out <- call_parent_method
       }
       out
-    }),
+    }
+  ),
   render_bg = function(panel_params, theme) {
-    panel_type <- eval(quote(self$layout[self$layout$PANEL==i,]$PANEL_TYPE), sys.parent(2))
+    panel_type <- eval(quote(self$layout[self$layout$PANEL == i, ]$PANEL_TYPE), sys.parent(2))
     if (is.element(panel_type, c("x", "y"))) {
       ggside_guide_grid(
         theme,
@@ -165,39 +175,48 @@ CoordSideTrans <- ggplot2::ggproto(
     }
   },
   render_fg = ggside_render_fg,
-  render_axis_h = function (panel_params, theme) {
+  render_axis_h = function(panel_params, theme) {
     panel_type <- panel_params$ggside_panel_type
-    if (panel_type=="y") {
-      list(top = ggside_panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
-           bottom = ggside_panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom))
+    if (panel_type == "y") {
+      list(
+        top = ggside_panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
+        bottom = ggside_panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom)
+      )
     } else {
-      list(top = panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
-           bottom = panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom))
+      list(
+        top = panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
+        bottom = panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom)
+      )
     }
   },
-  render_axis_v = function (panel_params, theme) {
+  render_axis_v = function(panel_params, theme) {
     panel_type <- panel_params$ggside_panel_type
-    if (panel_type=="x") {
-      list(left = ggside_panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
-           right = ggside_panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right))
+    if (panel_type == "x") {
+      list(
+        left = ggside_panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
+        right = ggside_panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right)
+      )
     } else {
-      list(left = panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
-           right = panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right))
+      list(
+        left = panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
+        right = panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right)
+      )
     }
   }
 )
 
 #' @rdname ggside-ggproto-coord
 #' @export
-ggside_coord.CoordFixed <- function(coord){
+ggside_coord.CoordFixed <- function(coord) {
   # insure classes that inherit from CoordCartesian fail
   # if there is no S3 method called.
   ggplot2::ggproto("CoordSide",
-                   CoordSideFixed,
-                   limits = coord$limits,
-                   ratio = coord$ratio,
-                   expand = coord$expand,
-                   clip = coord$clip)
+    CoordSideFixed,
+    limits = coord$limits,
+    ratio = coord$ratio,
+    expand = coord$expand,
+    clip = coord$clip
+  )
 }
 
 
@@ -207,14 +226,13 @@ CoordSideFixed <- ggplot2::ggproto(
   setup_panel_params = new_ggproto_fun(
     ggplot2::CoordFixed$setup_panel_params,
     {
-
       if (!is.null(old_y_lim <- self$limits$y) &
-          is.ggside_scale(scale_y)) {
+        is_ggside_scale(scale_y)) {
         self$limits$y <- NULL
         out <- call_parent_method
         self$limits$y <- old_y_lim
       } else if (!is.null(old_x_lim <- self$limits$x) &
-                 is.ggside_scale(scale_x)) {
+        is_ggside_scale(scale_x)) {
         self$limits$x <- NULL
         out <- call_parent_method
         self$limits$x <- old_x_lim
@@ -222,7 +240,8 @@ CoordSideFixed <- ggplot2::ggproto(
         out <- call_parent_method
       }
       out
-    }),
+    }
+  ),
   render_bg = function(panel_params, theme) {
     panel_type <- panel_params$ggside_panel_type
     if (is.element(panel_type, c("x", "y"))) {
@@ -245,25 +264,32 @@ CoordSideFixed <- ggplot2::ggproto(
     }
   },
   render_fg = ggside_render_fg,
-  render_axis_h = function (panel_params, theme) {
+  render_axis_h = function(panel_params, theme) {
     panel_type <- panel_params$ggside_panel_type
-    if (panel_type=="y") {
-      list(top = ggside_panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
-           bottom = ggside_panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom))
+    if (panel_type == "y") {
+      list(
+        top = ggside_panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
+        bottom = ggside_panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom)
+      )
     } else {
-      list(top = panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
-           bottom = panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom))
-
+      list(
+        top = panel_guides_grob(panel_params$guides, position = "top", theme = theme, labels = panel_params$draw_labels$top),
+        bottom = panel_guides_grob(panel_params$guides, position = "bottom", theme = theme, labels = panel_params$draw_labels$bottom)
+      )
     }
   },
-  render_axis_v = function (panel_params, theme) {
+  render_axis_v = function(panel_params, theme) {
     panel_type <- panel_params$ggside_panel_type
-    if (panel_type=="x") {
-      list(left = ggside_panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
-           right = ggside_panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right))
+    if (panel_type == "x") {
+      list(
+        left = ggside_panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
+        right = ggside_panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right)
+      )
     } else {
-      list(left = panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
-           right = panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right))
+      list(
+        left = panel_guides_grob(panel_params$guides, position = "left", theme = theme, labels = panel_params$draw_labels$left),
+        right = panel_guides_grob(panel_params$guides, position = "right", theme = theme, labels = panel_params$draw_labels$right)
+      )
     }
   }
 )