+++ NEWS.md
@@ -10,7 +10,7 @@ Below is a running log of changes that are required due to `ggplot2`'s welcome c
 * `ggplot2` still maintains some backward compatibility for R verions < 4.3, in that it will use their internal `add_gg()` function. This was always an internal function and I had to reimplement (and overload) as a hack to get `ggside` to behave correctly. Despite this, `ggside` will remove its own `add_gg()` function and solely rely on the S7 method dispatch for `+`. As a result, **`ggside > v0.3.1` will require R version >= 4.3.0** because we will no longer overload `+.gg` to maintain compatibility with older versions of R.
   * This decision is so that `ggside`'s code complexity can be reduced and lower the burden of maintaining the package.
 * Additionally `ggside` (> v0.3.1) will now depend on `ggplot2 (> v3.5.2.xxx)`.
-
+* Deprecating various `is.ggside*()` functions in favor of `is_ggside*()`.
 
 # ggside 0.3.2